pyproject.toml
[project]
name = "amazon-nova"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "litellm>=1.50.0",
    "python-dotenv>=1.0.0",
]
